requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
cloudinary==1.34.0
coreapi==2.3.3
coreschema==0.0.4
Cython==3.0.2
Django==4.2.4
django-ckeditor==6.7.0
django-cors-headers==4.2.0
django-js-asset==2.1.0
djangorestframework==3.14.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mysqlclient==2.2.0
packaging==23.1
python-decouple==3.8
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
